исправления репозитория и контроллера

+++ src/main/java/com/demo/controller/ProductController.java
@@ -1,7 +1,9 @@
 package com.demo.controller;
 
+import com.demo.model.Product;
 import com.demo.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
@@ -14,46 +16,55 @@ public class ProductController {
 
     // Создание продукта
     @PostMapping
-    public Product createUser(@RequestBody Product product) {
+    public Product createProduct(@RequestBody Product product) {
         return productRepository.save(product);
     }
     // Получение продукта по ID
     @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable Integer id) {
-        return productRepository.findById(id)
-                .map(product -> ResponseEntity.ok().body(product))
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<Product> getProductById(@PathVariable(name = "id") int id){
+        Product product = productRepository.findByIdProduct(id);
+        return product != null
+                ? new ResponseEntity<>(product, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
-    // Получение продукта по названию
-    @GetMapping("/{name}")
-    public ResponseEntity<Product> getProductByName(@PathVariable String name) {
-        return productRepository.findByName(name)
-                .map(product -> ResponseEntity.ok().body(product))
-                .orElse(ResponseEntity.notFound().build());
+    @GetMapping("/proteins/{startPro}/{endPro}")
+    public ResponseEntity<List<Product>> getProductsByProteins(@PathVariable float startPro, @PathVariable float endPro) {
+        List<Product> products = productRepository.findByProteinsBetween(startPro, endPro);
+        if (!products.isEmpty()) {
+            return ResponseEntity.ok().body(products);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
-    @GetMapping("/{proteins}")
-    public ResponseEntity<Product> getProductByProteins(@PathVariable float startPro, float endPro) {
-        return productRepository.findByProteinsBetween(startPro,endPro)
-                .map(product -> ResponseEntity.ok().body(product))
-                .orElse(ResponseEntity.notFound().build());
+
+    @GetMapping("/fats/{startFat}/{endFat}")
+    public ResponseEntity<List<Product>> getProductsByFats(@PathVariable float startFat, @PathVariable float endFat) {
+        List<Product> products = productRepository.findByFatsBetween(startFat, endFat);
+        if (!products.isEmpty()) {
+            return ResponseEntity.ok().body(products);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
-    @GetMapping("/{fats}")
-    public ResponseEntity<Product> getProductByFats(@PathVariable float startFat, float endFat) {
-        return productRepository.findByFatsBetween(startFat,endFat)
-                .map(product -> ResponseEntity.ok().body(product))
-                .orElse(ResponseEntity.notFound().build());
+
+    @GetMapping("/carbohydrates/{startCar}/{endCar}")
+    public ResponseEntity<List<Product>> getProductsByCarbohydrates(@PathVariable float startCar, @PathVariable float endCar) {
+        List<Product> products = productRepository.findByCarbohydratesBetween(startCar, endCar);
+        if (!products.isEmpty()) {
+            return ResponseEntity.ok().body(products);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
-    @GetMapping("/{carbohydrates}")
-    public ResponseEntity<Product> getProductByCarbohydrates(@PathVariable float startCar, float endCar) {
-        return productRepository.findByCarbohydratesBetween(startCar,endCar)
-                .map(product -> ResponseEntity.ok().body(product))
-                .orElse(ResponseEntity.notFound().build());
-    }
-    @GetMapping("/{calories}")
-    public ResponseEntity<Product> getProductByCarbohydrates(@PathVariable float startCal, float endCal) {
-        return productRepository.findByCaloriesBetween(startCal, endCal)
-                .map(product -> ResponseEntity.ok().body(product))
-                .orElse(ResponseEntity.notFound().build());
+
+    @GetMapping("/calories/{startCal}/{endCal}")
+    public ResponseEntity<List<Product>> getProductsByCalories(@PathVariable float startCal, @PathVariable float endCal) {
+        List<Product> products = productRepository.findByCaloriesBetween(startCal, endCal);
+        if (!products.isEmpty()) {
+            return ResponseEntity.ok().body(products);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
     // Получение всех продуктов
     @GetMapping
@@ -61,29 +72,34 @@ public class ProductController {
         return productRepository.findAll();
     }
 
-    // Обновление пользователя
+    // Обновление продукта
     @PutMapping("/{id}")
     public ResponseEntity<Product> updateProduct(@PathVariable Integer id, @RequestBody Product productDetails) {
-        return productRepository.findById(id)
-                .map(product -> {
-                    product.setName(productDetails.getName());
-                    product.setProteins(productDetails.getProteins());
-                    product.setFats(productDetails.getFats());
-                    product.setCarbohydrates(productDetails.getCarbohydrates());
-                    product.setCalories(productDetails.getCalories());
-                    Product updatedProduct = productRepository.save(product);
-                    return ResponseEntity.ok().body(updatedProduct);
-                })
-                .orElse(ResponseEntity.notFound().build());
+        Product product = productRepository.findByIdProduct(id);
+
+        if (product != null) {
+            product.setName(productDetails.getName());
+            product.setProteins(productDetails.getProteins());
+            product.setFats(productDetails.getFats());
+            product.setCarbohydrates(productDetails.getCarbohydrates());
+            product.setCalories(productDetails.getCalories());
+
+            Product updatedProduct = productRepository.save(product);
+            return ResponseEntity.ok().body(updatedProduct);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteProduct(@PathVariable Integer id) {
-        return productRepository.findById(id)
-                .map(product -> {
-                    productRepository.delete(product);
-                    return ResponseEntity.ok().<Void>build();
-                })
-                .orElse(ResponseEntity.notFound().build());
+        Product product = productRepository.findByIdProduct(id);
+
+        if (product != null) {
+            productRepository.delete(product);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
